requirements.txt
# Used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask>=0.11
apiai
pyYaml
SQLAlchemy>=1.1.11